[CoreBunde] User self registration.
