+++ main/core/Resources/modules/workspace/routing.js
@@ -1,6 +1,10 @@
 import {route as toolRoute} from '#/main/core/tool/routing'
 
 function route(workspace, toolName = null) {
+  if (!workspace) {
+    return null
+  }
+  
   if (toolName) {
     return toolRoute('workspaces')+`/open/${workspace.slug}/${toolName}`
   }
